pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.0"
python-dotenv = "^1.0.1"
tiktoken = "^0.7.0"
sentence-transformers = "^3.1.1"
langchain-community = "^0.3.0"
langchain-core = "^0.3.5"
faiss-cpu = "^1.8.0.post1"
gdown = "^5.2.0"
openpyxl = "^3.1.5"
pandas = "^2.2.3"
chainlit = ">1.3.2"
neo4j = "^5.26.0"
chromadb = "^0.5.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
